+++ typescript/registry/src/models/errors/index.ts
@@ -3,10 +3,9 @@
  */
 
 export * from "./apierror.js";
-export * from "./getserver.js";
+export * from "./error.js";
 export * from "./httpclienterrors.js";
 export * from "./responsevalidationerror.js";
 export * from "./sdkvalidationerror.js";
-export * from "./servererror.js";
 export * from "./smitheryregistryerror.js";
-export * from "./unauthorizederror.js";
+export * from "./uplinkerror.js";